requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

flask
requests
black

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
torch==1.9.0+cpu
torchvision==0.10.0+cpu
tqdm>=4.41.0

tensorboard>=2.4.1

seaborn>=0.11.0
pandas

gunicorn==20.1.0

thop  # FLOPs computation
